# docs/swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Complete AI infrastructure platform providing gateway, observability, caching, and optimization services.",
        "title": "Brokle AI Infrastructure Platform API",
        "termsOfService": "https://brokle.ai/terms",
        "contact": {
            "name": "Brokle Platform Team",
            "url": "https://brokle.ai/support",
            "email": "support@brokle.ai"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/api/v1/environments/{envId}/api-keys": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a paginated list of API keys for a specific environment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API Keys"
                ],
                "summary": "List API keys",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"env_1234567890\"",
                        "description": "Environment ID",
                        "name": "envId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "active",
                            "inactive",
                            "revoked"
                        ],
                        "type": "string",
                        "description": "Filter by API key status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of API keys",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/apikey.ListAPIKeysResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid environment ID",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - insufficient permissions to view API keys",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Environment not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new API key for an environment. The key will only be displayed once upon creation.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API Keys"
                ],
                "summary": "Create API key",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"env_1234567890\"",
                        "description": "Environment ID",
                        "name": "envId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "API key details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apikey.CreateAPIKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "API key created successfully (key only shown once)",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/apikey.APIKey"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid input or validation errors",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - insufficient permissions to create API keys",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Environment not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict - API key name already exists in environment",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/environments/{envId}/api-keys/{keyId}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Permanently revoke and delete an API key. This action cannot be undone and will immediately invalidate the key.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API Keys"
                ],
                "summary": "Delete API key",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"env_1234567890\"",
                        "description": "Environment ID",
                        "name": "envId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"key_1234567890\"",
                        "description": "API Key ID",
                        "name": "keyId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "API key deleted successfully"
                    },
                    "400": {
                        "description": "Bad request - invalid environment ID or key ID",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - insufficient permissions to delete API keys",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Environment or API key not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "apikey.APIKey": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "created_by": {
                    "type": "string",
                    "example": "usr_1234567890"
                },
                "environment_id": {
                    "type": "string",
                    "example": "env_1234567890"
                },
                "expires_at": {
                    "type": "string",
                    "example": "2024-12-31T23:59:59Z"
                },
                "id": {
                    "type": "string",
                    "example": "key_1234567890"
                },
                "key": {
                    "type": "string",
                    "example": "bk_live_1234567890abcdef"
                },
                "key_preview": {
                    "type": "string",
                    "example": "bk_live_...cdef"
                },
                "last_used": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "name": {
                    "type": "string",
                    "example": "Production API Key"
                },
                "scopes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"read\"",
                        " \"write\"]"
                    ]
                },
                "status": {
                    "type": "string",
                    "example": "active"
                }
            }
        },
        "apikey.CreateAPIKeyRequest": {
            "type": "object",
            "required": [
                "environment_id",
                "name",
                "scopes"
            ],
            "properties": {
                "environment_id": {
                    "type": "string",
                    "example": "env_1234567890"
                },
                "expires_at": {
                    "type": "string",
                    "example": "2024-12-31T23:59:59Z"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2,
                    "example": "Production API Key"
                },
                "scopes": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"read\"",
                        " \"write\"]"
                    ]
                }
            }
        },
        "apikey.ListAPIKeysResponse": {
            "type": "object",
            "properties": {
                "api_keys": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/apikey.APIKey"
                    }
                },
                "limit": {
                    "type": "integer",
                    "example": 20
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "total": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
...
        "response.APIError": {
            "description": "Error details for failed API requests",
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "validation_error"
                },
                "details": {
                    "type": "string",
                    "example": "Field 'email' is required"
                },
                "message": {
                    "type": "string",
                    "example": "Invalid request data"
                },
                "type": {
                    "type": "string",
                    "example": "validation_error"
                }
            }
        },
        "response.ErrorResponse": {
            "description": "Standard error response",
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/response.APIError"
                },
                "meta": {
                    "$ref": "#/definitions/response.Meta"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "response.Meta": {
            "description": "Response metadata including request tracking and pagination",
            "type": "object",
            "properties": {
                "pagination": {
                    "$ref": "#/definitions/response.Pagination"
                },
                "request_id": {
                    "type": "string",
                    "example": "req_01h2x3y4z5"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2023-12-01T10:30:00Z"
                },
                "total": {
                    "type": "integer",
                    "example": 150
                },
                "version": {
                    "type": "string",
                    "example": "v1"
                }
            }
        },
        "response.Pagination": {
            "description": "Pagination metadata for paginated API responses",
            "type": "object",
            "properties": {
                "has_next": {
                    "type": "boolean",
                    "example": true
                },
                "has_prev": {
                    "type": "boolean",
                    "example": false
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "page_size": {
                    "type": "integer",
                    "example": 20
                },
                "total": {
                    "type": "integer",
                    "example": 150
                },
                "total_page": {
                    "type": "integer",
                    "example": 8
                }
            }
        },
        "response.SuccessResponse": {
            "description": "Standard successful response",
            "type": "object",
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/response.Meta"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API key for authentication. Format: X-API-Key: bk_live_...",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "JWT token for authentication. Format: Authorization: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}